Various fixes after merge of stencil
