+++ rust-app/src/handle_apdu.rs
@@ -1,8 +1,6 @@
 use crate::implementation::*;
 use crate::interface::*;
 use crate::settings::*;
-#[cfg(not(any(target_os = "stax", target_os = "flex")))]
-use crate::test_parsers::*;
 use crate::utils::*;
 
 use alamgu_async_block::*;
