+++ rust-app/src/ui/nano.rs
@@ -1,18 +1,42 @@
+use crate::interface::*;
+use crate::utils::*;
 use core::fmt::Write;
-use ledger_crypto_helpers::eddsa::Ed25519RawPubKeyAddress;
-use ledger_crypto_helpers::hasher::Base64Hash;
+use ledger_crypto_helpers::common::HexSlice;
+use ledger_crypto_helpers::hasher::HexHash;
 use ledger_prompts_ui::*;
 
-pub fn confirm_address(pkh: &Ed25519RawPubKeyAddress) -> Option<()> {
+pub fn confirm_address(address: &SuiPubKeyAddress) -> Option<()> {
     scroller("Provide Public Key", |_w| Ok(()))?;
-    scroller_paginated("Address", |w| Ok(write!(w, "{pkh}")?))?;
+    scroller_paginated("Address", |w| Ok(write!(w, "{address}")?))?;
     final_accept_prompt(&[])
 }
 
-pub fn confirm_sign_tx(pkh: &Ed25519RawPubKeyAddress, hash: &Base64Hash<32>) -> Option<()> {
+pub fn confirm_sign_tx(
+    address: &SuiPubKeyAddress,
+    recipient: [u8; 32],
+    total_amount: u64,
+    gas_budget: u64,
+) -> Option<()> {
+    scroller("Transfer", |w| Ok(write!(w, "SUI")?))?;
+
+    scroller_paginated("From", |w| Ok(write!(w, "{address}")?))?;
+    scroller_paginated("To", |w| Ok(write!(w, "0x{}", HexSlice(&recipient))?))?;
+
+    let (quotient, remainder_str) = get_amount_in_decimals(total_amount);
+    scroller_paginated("Amount", |w| {
+        Ok(write!(w, "SUI {quotient}.{}", remainder_str.as_str())?)
+    })?;
+
+    let (quotient, remainder_str) = get_amount_in_decimals(gas_budget);
+    scroller("Max Gas", |w| {
+        Ok(write!(w, "SUI {}.{}", quotient, remainder_str.as_str())?)
+    })?;
+    final_accept_prompt(&["Sign Transaction?"])
+}
+
+pub fn confirm_blind_sign_tx(hash: &HexHash<32>) -> Option<()> {
     scroller("WARNING", |w| Ok(write!(w, "Transaction not recognized")?))?;
-    scroller("Transaction hash", |w| Ok(write!(w, "{}", hash)?))?;
-    scroller("Sign for Address", |w| Ok(write!(w, "{pkh}")?))?;
+    scroller("Transaction Hash", |w| Ok(write!(w, "0x{hash}")?))?;
     final_accept_prompt(&["Blind Sign Transaction?"])
 }
